requirements.txt
pyyaml==6.0.1
numpy==1.26.1
Pillow==10.0.1
scikit-image==0.22.0
opencv-python==4.8.1.78